+++ Server/src/app.js
@@ -3,7 +3,7 @@ const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 const morgan = require('morgan');
 const routes = require('./routes/index');
-const routerCloud = require('./cloudinary/imgPetsRoutes.js');
+const router = require('./routes/imagePetsRoutes.js');
 
 require('./db.js'); //conexion con la DB
 
@@ -23,10 +23,9 @@ server.use((req, res, next) => {
   next();
 });
 
+server.use('/api/pets', router); //rutas para cargar y pedir imágenes de las mascotas a cloudinary  
 
 server.use('/', routes) //raiz del enrutamiento Back
-
-server.use('/api/pets', routerCloud); //rutas para cargar y pedir imágenes de las mascotas a cloudinary 
 
 //midleware para manejo de eventuales errores en las rutas
 server.use((err, req, res, next) => {
@@ -35,6 +34,6 @@ server.use((err, req, res, next) => {
     console.error(err);
     res.status(status).send(message);
   });
-  
+
 
 module.exports = server
