+++ Server/src/routes/index.js
@@ -8,7 +8,7 @@ const routerPets = require('./routesPets')
 
 // const routerFilter = require('./routesFilter')
 const routerForUsers = require('./routesUser.js') //----Marcos
-const routerCloud = require('../cloudinary/imgPetsRoutes')
+
 // const {} = require('./routesDonations')
 // const {} = require('./routesAdoptions')
 
@@ -19,7 +19,7 @@ routes.use('/mascotas', routerPets)  // Model : mascota
 
 // routes.use('/filter', routerFilter)
 routes.use('/usuario', routerForUsers) // Model : User ----Marcos
-routes.use('/cloud', routerCloud) // Ruta para la carga de imágenes de las mascotas en Cloudinary
+
 
 // routes.use('/donacion', /*rutaDonacion*/) // Model : Donacion
 // routes.use('/Adopcion', /*rutaAdopcion*/) // Model : Adopcion
